+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to **Sessions** are documented in this *changelog*.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and **Sessions** adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased - will be 1.1.1]
+### Changed
+- Expired sessions cookies are now counted as cleaned sessions.
 ### Fixed
 - The counted deleted user may be wrong in KPIs.
 - Batch sessions deletion are wrongly counted.